Fixes and connected solution
There were some fixes to the GameObject and Component classes so that they work properly.

The game solution has been linked to the engine solution.

--- Work in progress ---
The InputManager and the Command classes have been added.
InputManager supports multiple Xbox controllers and can check if a button has been pressed or released.