Ingredient and Plate
First draft of the Ingredient class (controls the ingredients of the burger, moves them, checks if it's overlapped by enemies, etc.).

First draft of the Plate class, the plate is the final destination of the burger. Checks if a full burger is on it, resets the plate on a new level and adds ingredients when they reach the plate/the burger.